Move server startup to rootcmd

+++ common/common.go
@@ -18,7 +18,7 @@ type Options struct {
 	BindAddr      string `json:"bind_address,omitempty"`
 	TLSCertPath   string `json:"tls_cert_path,omitempty"`
 	TLSKeyPath    string `json:"tls_cert_key,omitempty"`
-	LogLevel      int    `json:"log_level,omitempty"`
+	LogLevel      uint64 `json:"log_level,omitempty"`
 	LogFile       string `json:"log_file,omitempty"`
 	ZcashConfPath string `json:"zcash_conf,omitempty"`
 	VeryInsecure  bool   `json:"very_insecure,omitempty"`
